noob int

+++ include/agent.h
@@ -26,7 +26,7 @@ int AGENT_move(AGENT *agent, MATRIX *world, int action);
 
 int AGENT_unmove(AGENT *agent, MATRIX *world, int action);
 
-int choose_action(AGENT *agent, MATRIX *world, int alfa, float default_value);
+int choose_action(AGENT *agent, MATRIX *world, float alfa, float default_value);
 
 int AGENT_change_pos(AGENT *agent, MATRIX *world, int newx, int newy);
 
@@ -34,7 +34,7 @@ int AGENT_is_wall(MATRIX *world, int newx, int newy);
 
 int choose_best_action(AGENT *agent, MATRIX *world, float default_value);
 
-void Q_learning(AGENT *agent, MATRIX *world, int alfa, int gamma, float default_value);
+void Q_learning(AGENT *agent, MATRIX *world, float alfa, float gamma, float default_value);
 
 void AGENT_reset(AGENT *agent);
 #endif
